Change to the currently preferred "two-file explicit instantiation" scheme.
