+++ drake/automotive/BUILD
@@ -79,7 +79,6 @@ drake_cc_library(
     srcs = ["endless_road_car.cc"],
     hdrs = [
         "endless_road_car.h",
-        "endless_road_car-inl.h",
         "endless_road_car_to_euler_floating_joint.h",
     ],
     linkstatic = 1,
@@ -94,7 +93,7 @@ drake_cc_library(
 drake_cc_library(
     name = "endless_road_oracle",
     srcs = ["endless_road_oracle.cc"],
-    hdrs = ["endless_road_oracle.h", "endless_road_oracle-inl.h"],
+    hdrs = ["endless_road_oracle.h"],
     linkstatic = 1,
     deps = [
         ":generated_vectors",
